rust与mysql类型匹配对应

+++ web-pm-rust/src/entity/accounts.rs
@@ -9,14 +9,14 @@ use crate::util::option_date_format;
 #[serde(rename_all = "camelCase", default)]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: u64,
     pub act_introduction: Option<String>,
     pub act_name: String,
     pub act_nick_name: Option<String>,
     pub act_pwd: String,
     #[serde(with = "option_date_format")]
     pub act_register_date: Option<DateTime>,
-    pub act_status: Option<i8>,
+    pub act_status: Option<u8>,
     #[serde(with = "option_date_format")]
     pub created_at: Option<DateTime>,
     #[serde(with = "option_date_format")]
@@ -35,11 +35,11 @@ impl Model {
         Entity::find().all(db).await.unwrap()
     }
 
-    pub async fn find_by_id(db: &DbConn, id: i32) -> Option<Model> {
+    pub async fn find_by_id(db: &DbConn, id: u64) -> Option<Model> {
         Entity::find_by_id(id).one(db).await.unwrap()
     }
 
-    pub async fn save(db: &DbConn, account: Model) -> i32 {
+    pub async fn save(db: &DbConn, account: Model) -> u64 {
         let insert_account: ActiveModel = account.into_active_model();
         let insert_result = Entity::insert(insert_account).exec(db).await.unwrap();
         insert_result.last_insert_id
